# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Список",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Фен для волос"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "fen-dlya-volos"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 937.4
                                                    },
                                                    "external_id": {
                                                        "type": "integer",
                                                        "example": 3241
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Создание",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Фен для волос"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "fen-dlya-volos"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 937.4
                                            },
                                            "external_id": {
                                                "type": "integer",
                                                "example": 3241
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Фен для волос"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "fen-dlya-volos"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 937.4
                                                },
                                                "external_id": {
                                                    "type": "integer",
                                                    "example": 3241
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{slug}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Единичная запись",
                "operationId": "5bbb6d725346d38781b256ed5c715d62",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug продукта",
                        "required": true,
                        "example": "fen-dlya-volos"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Обновление",
                "operationId": "623bc63ad26c7e8228a8fc83b7362908",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug продукта",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Фен для волос изменить"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1937.5
                                            },
                                            "external_id": {
                                                "type": "integer",
                                                "example": 3241
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Фен для волос"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "fen-dlya-volos"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 937.4
                                                },
                                                "external_id": {
                                                    "type": "integer",
                                                    "example": 3241
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Удаление",
                "operationId": "7a1a68407fb59d2cb9e9c5d646ea7d73",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug Продукта",
                        "required": true,
                        "example": "fen-dlya-volos"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}